Wait for host to be offline before reconnecting in test_reboot_server
When triggering a reboot using `systemctl reboot`, it is possible for the SSH
server to remain active for a short time. This causes a race-condition with our
reconnection step.

Though this issue was discovered with Debian 11, which seems to have the longest
delay of all tested images, it is an issue that could occur on any other image.