Limit generated server names to 63 characters
The server name is limited to 63 characters when creating a server. The
name is simply truncated at the end. This improves the server name
generation to not create names longer than 63 characters, but keeps the
user supplied name at the end untruncated. This ensures that servers
with different name parameters are not created with the same name and
that they can be identified by their names.