Replace retry_for by wait_for_X_ready functions
This should make the test logic a bit more explicit by not mixing the
test to wait for the configuration change to be active on the load
balancer with the actual test assertion. On the expense that the
test assertion might be redundant as it does not assert anything new.