Improve robustness of two_servers_in_same_subnet
The previous approached relied on launching enough servers in parallel
to hopefully get two in the same subnet. This got more unreliablie
as we added more public subnets.

Instead of launching even more servers in parallel, we now create
servers serially, but throw them away early, if they are not a match.

This relies on the fact that the API does not delete the servers
synchronously (we otherwise might get the same subnet over and over
again). But that is unlikely to change and it is not more reliant on
internal behavior than the previous solution.

In practice, this seems to work slightly faster. Before, with six
test runs and a function to check how long it takes to create the
fixture:

  Found two servers in 60.2s
  Found two servers in 56.3s
  Found two servers in 55.7s
  Found two servers in 55.4s
  Found two servers in 57.7s
  Found two servers in 57.4s

After:

  Found two servers in 40.7s
  Found two servers in 39.9s
  Found two servers in 39.7s
  Found two servers in 40.0s
  Found two servers in 39.8s
  Found two servers in 43.2s