Add the ability to manually run a subset of tests
This makes it easier to only run certain tests, or to develop changes
to the workflows without having to run the whole gamut.

Additionally, it is now possible to test workflow runs, without
impacting the result of our scheduled tests (which otherwise would
influence the badge in our README).

Each workflow run now has a separate token, which ensures that running
acceptance tests concurrently does not cause overlapping cleanups.

This also unifies the way we run tests and it is now easier to change
the way tests are run, without having to copy&paste code between
workflows.