Use a different pytest rerun approach
Instead of rerunning tests right after they failed, we now rerun them
at the end of the test-run. This is meant to deal with short-term
flakiness, and running the tests further apart should help.

Additionally, this now always outputs errors, even if its one that
was then "fixed" in a rerun. This way we can check the causes
for reruns.