Load balancer failover test
Verify basic load balancer operation after a failover. During the
failover load balancer VMs will be rebuilt. This test verifies that the
new VMs work as expected.

This is an interactive test as load balancer failover is an
administrative task which can only be executed by cloudscale.ch staff.
This operation is not available on the public API.