Change multiple snapshots test to be more generic
I felt like the previous approach was, though correct and a good idea,
a bit hard to follow with the number of snapshots and repeated code.

Though we generally do prefer repeated code over loops in our acceptance
tests, I think here it helps clarify the test:

1. In a first block, snapshots are created with evidence.
2. In a second block, snapshots re restored, and evidence is gathered.
3. In the final block, evidence is compared.

To me, especially the last block is much more obvious, as it is now
quite clear how the evidence changes over time.