Log request body for failed requests
Log the request body to the event log for failed requests. This often
contains a more detailed error message. Successful request bodies are
not logged. They may contain sensitive information and are generally
much larger.

This removes the "print" logging which was done before. This was the
only part of the code which logged using print to standard output.